requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
openai-whisper==20231117
transformers==4.35.2
torch==2.1.1
TTS==0.22.0
ffmpeg-python==0.2.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==0.8.0
aiofiles==0.8.0
moviepy==1.0.3
numpy==1.26.3
pydub==0.25.1
opencv-python==4.9.0.80
scipy==1.12.0
soundfile==0.12.1
python-jose[cryptography]==3.3.0
redis==5.0.1
celery==5.3.6
flower==2.0.1
face-alignment==1.3.5 